requirements.txt
Keras==2.2.5
keras-resnet==0.2.0
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==6.2.0
tensorflow==1.14.0
progressbar2
Flask-Cors
Flask
psutil==5.7.0
Werkzeug
shapely==1.7.0
pyclipper
tesserocr